django_api/requirements/sub-master.txt
Django==1.11
  - pytz [required: Any, installed: 2018.3]
djangorestframework==3.7.7
pipdeptree==0.11.0
  - pip [required: >=6.0.0, installed: 9.0.1]
setuptools==38.4.0
wheel==0.30.0
